requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
protobuf==3.9.1
six==1.12.0
Werkzeug==0.15.5
